TrainDesign (RAS 2011): data files + begin of the importer + unneeded long casting

+++ drools-planner-examples/src/main/java/org/drools/planner/examples/manners2009/persistence/Manners2009SolutionImporter.java
@@ -165,7 +165,7 @@ public class Manners2009SolutionImporter extends AbstractTxtSolutionImporter {
             long id = 0L;
             for (Guest guest : guestList) {
                 SeatDesignation seatDesignation = new SeatDesignation();
-                seatDesignation.setId((long) id);
+                seatDesignation.setId(id);
                 id++;
                 seatDesignation.setGuest(guest);
                 // Notice that we leave the PlanningVariable properties on null
